backend/python/similarity_api/requirements.txt
psycopg2-binary
langchain
langchain_google_genai
pinecone
